#46 Added `GATTDescriptorTests`
